move the combined protocol schedule into its own class
